Split looper state into its own file
